move hosts resolving to SeedHostsResolver
